Move cli/util intercept related stuff to cli/intercept package.
