+++ pkg/client/cli/intercept/describe_intercepts.go
@@ -1,4 +1,4 @@
-package util
+package intercept
 
 import (
 	"fmt"
