+++ pkg/client/cli/intercept/prepare_mount_windows.go
@@ -1,4 +1,4 @@
-package util
+package intercept
 
 import (
 	"fmt"
